Feature/logicDelete

Approved-by: Alejandro García Granados

+++ micronaut/finerio-pfm-api/src/main/groovy/mx/finerio/pfm/api/domain/Account.groovy
@@ -18,7 +18,7 @@ class Account  implements GormEntity<Account> {
     String nature
     String name
     String number
-    BigDecimal balance
+    float balance
     Date dateCreated
     Date lastUpdated
     Date dateDeleted
